refactor(Cosmology): add type annotations to methods
